+++ lib/widgets/ny_page.dart
@@ -1,6 +1,7 @@
 import 'package:skeletonizer/skeletonizer.dart';
 
-import '../event_bus/res/history_entry.dart';
+import '/helpers/loading_style.dart';
+import '/event_bus/res/history_entry.dart';
 import '/nylo.dart';
 import '/widgets/ny_base_state.dart';
 import '/router/models/nyrouter_route_guard.dart';